requirements.txt
fastapi
uvicorn
aiofiles
pillow
torch
transformers